pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discord-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for Discord"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.0",
]

[project.scripts]
discord-mcp = "src.main:run"

# Project name is discord-mcp; package dir is `src/` (imports: from src.tools …)
[tool.hatch.build.targets.wheel]
packages = ["src"]